Replace float with double

+++ src/xml2db/table/column.py
@@ -3,7 +3,7 @@ from typing import List, Iterable, Any, Union, TYPE_CHECKING
 
 from sqlalchemy import (
     Integer,
-    Float,
+    Double,
     Boolean,
     BigInteger,
     SmallInteger,
@@ -30,7 +30,7 @@ def types_mapping_default(temp: bool, col: "DataModelColumn") -> Any:
     if col.occurs[1] != 1:
         return String(8000)
     if col.data_type in ["decimal", "float"]:
-        return Float
+        return Double
     if col.data_type == "dateTime":
         return DateTime(timezone=True)
     if col.data_type == "integer" or col.data_type == "int":
@@ -72,7 +72,7 @@ def types_mapping_mssql(temp: bool, col: "DataModelColumn") -> Any:
     if col.occurs[1] != 1:
         return mssql.VARCHAR(8000)
     if col.data_type in ["decimal", "float"]:
-        return Float
+        return Double
     if col.data_type == "dateTime":
         # using the DATETIMEOFFSET directly in the temporary table caused issues when inserting data in the target
         # table with INSERT INTO SELECT converts datetime VARCHAR to DATETIMEOFFSET without errors
